Add data dictionaries and code to autotabulate FPI results for a branch
The remaining issue is that the FPI results are not particularly informative
with the correlation cutoff at 0.5.  So, it will be better to re-run FPI with
correlation cutoff at 0.7.  However, instead of rerunning all the ML all over
again (and generating new models, which are not exactly the same), let's create
an FPI-fine-tuning notebook that faciliates rerunning FPI at different cutoffs.
This does not require retraining and provides an example for how to use an
archived ML model. This will be addressed in the next commit.